WIP Section 4.1.2
Add application helper to provide a full title for each page. In
the event a title is not available the helper will provide a
default title.

Add test to check for the condition where no title is available.
Test is red at this stage.

Remove the call to provide function in home.html.erb. This gets the
test to green.