Complete section 8.2.2 and 8.2.3
Retrieve user id from current session in order to use on subsequent
pages.

Update layout links to give visual indication of whether a user is
logged in or not.

For logged in users, show an account link in the form of a dropdown.
We need some javascript libraries required by bootstrap to enable this.