11.2.4 Updating the Users create action
Update the create action in the users controller to respond properly
now that new users must be activated before saving to the databse.

11.2.4 Updating the Users create action

Update the create action in the users controller to respond properly
now that new users must be activated before saving to the databse.

11.2.4 Updating the Users create action

Update the create action in the users controller to respond properly
now that new users must be activated before saving to the databse.

11.2.4 Updating the Users create action

Update the create action in the users controller to respond properly
now that new users must be activated before saving to the databse.

11.2.4 Updating the Users create action

Update the create action in the users controller to respond properly
now that new users must be activated before saving to the databse.

11.2.4 Updating the Users create action

Update the create action in the users controller to respond properly
now that new users must be activated before saving to the databse.

11.2.4 Updating the Users create action

Update the create action in the users controller to respond properly
now that new users must be activated before saving to the databse.

11.2.4 Updating the Users create action

Update the create action in the users controller to respond properly
now that new users must be activated before saving to the databse.

11.2.4 Updating the Users create action

Update the create action in the users controller to respond properly
now that new users must be activated before saving to the databse.

11.2.4 Updating the Users create action

Update the create action in the users controller to respond properly
now that new users must be activated before saving to the databse.

11.2.4 Updating the Users create action

Update the create action in the users controller to respond properly
now that new users must be activated before saving to the databse.