14.2.4 Working follow button the standard way
Add conroller with create and delete actions to support following
and unfollowing users.

Add tests and fix test data for relationships. This had broken a
test in user_test.rb

Update Rails, bootsnap and puma gems per reference repo. This was
needed due to a failure when running the test suite after creating
the relationships controller.